+++ client/browser/src/shared/code-hosts/shared/codeHost.tsx
@@ -37,7 +37,6 @@ import {
     mapTo,
     take,
 } from 'rxjs/operators'
-import { HoverAlert } from 'sourcegraph'
 
 import { HoverMerged } from '@sourcegraph/client-api'
 import {
@@ -47,7 +46,6 @@ import {
     Hoverifier,
     HoverState,
     MaybeLoadingResult,
-    DiffPart,
 } from '@sourcegraph/codeintellify'
 import {
     asError,
@@ -60,13 +58,13 @@ import {
     LineOrPositionOrRange,
     lprToSelectionsZeroIndexed,
 } from '@sourcegraph/common'
-import { TextDocumentDecoration, WorkspaceRoot } from '@sourcegraph/extension-api-types'
+import { WorkspaceRoot } from '@sourcegraph/extension-api-types'
 import { gql, isHTTPAuthError } from '@sourcegraph/http-client'
 import { ActionItemAction, urlForClientCommandOpen } from '@sourcegraph/shared/src/actions/ActionItem'
 import { wrapRemoteObservable } from '@sourcegraph/shared/src/api/client/api/common'
-import { DecorationMapByLine } from '@sourcegraph/shared/src/api/extension/api/decorations'
 import { CodeEditorData, CodeEditorWithPartialModel } from '@sourcegraph/shared/src/api/viewerTypes'
 import { isRepoNotFoundErrorLike } from '@sourcegraph/shared/src/backend/errors'
+import { HoverAlert } from '@sourcegraph/shared/src/codeintel/legacy-extensions/api'
 import {
     CommandListClassProps,
     CommandListPopoverButtonClassProps,
@@ -116,7 +114,7 @@ import { phabricatorCodeHost } from '../phabricator/codeHost'
 
 import { CodeView, trackCodeViews, fetchFileContentForDiffOrFileInfo } from './codeViews'
 import { NotAuthenticatedError, RepoURLParseError } from './errors'
-import { applyDecorations, initializeExtensions, renderCommandPalette, renderGlobalDebug } from './extensions'
+import { initializeExtensions, renderCommandPalette } from './extensions'
 import { createRepoNotFoundHoverAlert, getActiveHoverAlerts, onHoverAlertDismissed } from './hoverAlerts'
 import {
     handleNativeTooltips,
@@ -324,11 +322,10 @@ export interface FileInfoWithContent extends FileInfoWithRepoName {
 export interface CodeIntelligenceProps extends TelemetryProps {
     platformContext: Pick<
         BrowserPlatformContext,
-        'urlToFile' | 'sideloadedExtensionURL' | 'requestGraphQL' | 'settings' | 'refreshSettings' | 'sourcegraphURL'
+        'urlToFile' | 'requestGraphQL' | 'settings' | 'refreshSettings' | 'sourcegraphURL'
     >
     codeHost: CodeHost
     extensionsController: Controller
-    showGlobalDebug?: boolean
 }
 
 export const getExistingOrCreateOverlayMount = (codeHostName: string, container: HTMLElement): HTMLElement => {
@@ -340,13 +337,6 @@ export const getExistingOrCreateOverlayMount = (codeHostName: string, container:
         container.append(mount)
     }
 
-    return mount
-}
-
-export const createGlobalDebugMount = (): HTMLElement => {
-    const mount = document.createElement('div')
-    mount.dataset.globalDebug = 'true'
-    document.body.append(mount)
     return mount
 }
 
@@ -822,7 +812,6 @@ export async function handleCodeHost({
     codeHost,
     extensionsController,
     platformContext,
-    showGlobalDebug,
     telemetryService,
     render,
     minimalUI,
@@ -933,14 +922,6 @@ export async function handleCodeHost({
                 })
             )
         )
-    }
-
-    // Render extension debug menu
-    // This renders to document.body, which we can assume is never removed,
-    // so we don't need to subscribe to mutations.
-    if (showGlobalDebug && extensionsController !== null) {
-        const mount = createGlobalDebugMount()
-        renderGlobalDebug({ extensionsController, platformContext, history, sourcegraphURL, render })(mount)
     }
 
     const signInCloses = new Subject<void>()
@@ -1273,7 +1254,6 @@ export async function handleCodeHost({
                         // For diffs, both editors are created (for head and base)
                         // but only one of them is passed into
                         // the `scope` of the CodeViewToolbar component.
-                        // Both are used to listen for text decorations.
                         const [headEditor, baseEditor] = await Promise.all([
                             initializeModelAndViewerForFileInfo(diffOrFileInfo.head),
                             initializeModelAndViewerForFileInfo(diffOrFileInfo.base),
@@ -1334,62 +1314,6 @@ export async function handleCodeHost({
                         target.closest('.sourcegraph-extension-element') !== null
                             ? null
                             : codeViewEvent.dom.getCodeElementFromTarget(target),
-                }
-
-                const applyDecorationsForFileInfo = (editor: CodeEditorWithPartialModel, diffPart?: DiffPart): void => {
-                    let decorationsByLine: DecorationMapByLine = new Map()
-                    let previousIsLightTheme = true
-                    const update = (decorations?: TextDocumentDecoration[] | null, isLightTheme?: boolean): void => {
-                        try {
-                            decorationsByLine = applyDecorations(
-                                domFunctions,
-                                element,
-                                decorations ?? [],
-                                decorationsByLine,
-                                isLightTheme ?? true,
-                                previousIsLightTheme,
-                                diffPart
-                            )
-                            previousIsLightTheme = isLightTheme ?? true
-                        } catch (error) {
-                            console.error('Could not apply decorations to code view', codeViewEvent.element, error)
-                        }
-                    }
-
-                    codeViewEvent.subscriptions.add(
-                        combineLatest([
-                            from(extensionsController.extHostAPI).pipe(
-                                switchMap(extensionHostAPI =>
-                                    wrapRemoteObservable(
-                                        extensionHostAPI.getTextDecorations({
-                                            viewerId: editor.viewerId,
-                                        })
-                                    )
-                                )
-                            ),
-                            codeHost.isLightTheme ?? of(true),
-                        ])
-                            // Make sure extensions get cleaned up un unsubscription
-                            .pipe(finalize(update))
-                            // The nested subscribe cannot be replaced with a switchMap()
-                            // We manage the subscription correctly.
-                            // eslint-disable-next-line rxjs/no-nested-subscribe
-                            .subscribe(([decorations, isLightTheme]) => update(decorations, isLightTheme))
-                    )
-                }
-
-                // Apply decorations coming from extensions
-                if (!minimalUI) {
-                    if ('blob' in diffOrFileInfoWithEditor) {
-                        applyDecorationsForFileInfo(diffOrFileInfoWithEditor.blob.editor)
-                    } else {
-                        if (diffOrFileInfoWithEditor.head) {
-                            applyDecorationsForFileInfo(diffOrFileInfoWithEditor.head.editor, 'head')
-                        }
-                        if (diffOrFileInfoWithEditor.base) {
-                            applyDecorationsForFileInfo(diffOrFileInfoWithEditor.base.editor, 'base')
-                        }
-                    }
                 }
 
                 // Add hover code navigation
@@ -1472,8 +1396,6 @@ export async function handleCodeHost({
     return subscriptions
 }
 
-const SHOW_DEBUG = (): boolean => localStorage.getItem('debug') !== null
-
 const CODE_HOSTS: CodeHost[] = [
     bitbucketServerCodeHost,
     bitbucketCloudCodeHost,
@@ -1547,8 +1469,7 @@ export function injectCodeIntelligenceToCodeHost(
     mutations: Observable<MutationRecordLike[]>,
     codeHost: CodeHost,
     { sourcegraphURL, assetsURL }: SourcegraphIntegrationURLs,
-    isExtension: boolean,
-    showGlobalDebug = SHOW_DEBUG()
+    isExtension: boolean
 ): Subscription {
     const subscriptions = new Subscription()
     const { platformContext, extensionsController } = initializeExtensions(
@@ -1617,7 +1538,6 @@ export function injectCodeIntelligenceToCodeHost(
                     codeHost,
                     extensionsController,
                     platformContext,
-                    showGlobalDebug,
                     telemetryService,
                     render: renderWithThemeProvider as Renderer,
                     minimalUI,