+++ client/extension-api/src/sourcegraph.d.ts
@@ -544,91 +544,6 @@ declare module 'sourcegraph' {
      */
     export type ViewComponent = CodeEditor | DirectoryViewer
 
-    /**
-     * A style for a {@link TextDocumentDecoration}.
-     */
-    export interface ThemableDecorationStyle {
-        /** The CSS background-color property value for the line. */
-        backgroundColor?: string
-
-        /** The CSS border property value for the line. */
-        border?: string
-
-        /** The CSS border-color property value for the line. */
-        borderColor?: string
-
-        /** The CSS border-width property value for the line. */
-        borderWidth?: string
-    }
-
-    /**
-     * A text document decoration changes the appearance of a range in the document and/or adds other content to
-     * it.
-     */
-    export interface TextDocumentDecoration extends ThemableDecorationStyle {
-        /**
-         * The range that the decoration applies to. Currently, decorations are
-         * only applied only on the start line, and the entire line. Multiline
-         * and intra-line ranges are not supported.
-         */
-        range: Range
-
-        /**
-         * If true, the decoration applies to all lines in the range (inclusive), even if not all characters on the
-         * line are included.
-         */
-        isWholeLine?: boolean
-
-        /** Content to display after the range. */
-        after?: DecorationAttachmentRenderOptions
-
-        /** Overwrite style for light themes. */
-        light?: ThemableDecorationStyle
-
-        /** Overwrite style for dark themes. */
-        dark?: ThemableDecorationStyle
-    }
-
-    /**
-     * A style for {@link DecorationAttachmentRenderOptions}.
-     */
-    export interface ThemableDecorationAttachmentStyle {
-        /** The CSS background-color property value for the attachment. */
-        backgroundColor?: string
-
-        /** The CSS color property value for the attachment. */
-        color?: string
-    }
-
-    /** A decoration attachment adds content after a {@link TextDocumentDecoration}. */
-    export interface DecorationAttachmentRenderOptions extends ThemableDecorationAttachmentStyle {
-        /** Text to display in the attachment. */
-        contentText?: string
-
-        /** Tooltip text to display when hovering over the attachment. */
-        hoverMessage?: string
-
-        /** If set, the attachment becomes a link with this destination URL. */
-        linkURL?: string
-
-        /** Overwrite style for light themes. */
-        light?: ThemableDecorationAttachmentStyle
-
-        /** Overwrite style for dark themes. */
-        dark?: ThemableDecorationAttachmentStyle
-    }
-
-    /**
-     * Represents a handle to a set of decorations.
-     *
-     * To get an instance of {@link TextDocumentDecorationType}, use
-     * {@link sourcegraph.app.createDecorationType}
-     */
-    export interface TextDocumentDecorationType {
-        /** An opaque identifier. */
-        readonly key: string
-    }
-
     export interface Directory {
         /**
          * The URI of the directory.
@@ -688,16 +603,6 @@ declare module 'sourcegraph' {
          * is always at index 0 of the emitted array.
          */
         readonly selectionsChanges: Subscribable<Selection[]>
-
-        /**
-         * Add a set of decorations to this editor. If a set of decorations already exists with the given
-         * {@link TextDocumentDecorationType}, they will be replaced.
-         *
-         * @see {@link TextDocumentDecorationType}
-         * @see {@link sourcegraph.app.createDecorationType}
-         *
-         */
-        setDecorations(decorationType: TextDocumentDecorationType, decorations: TextDocumentDecoration[]): void
     }
 
     /**
@@ -939,134 +844,6 @@ declare module 'sourcegraph' {
         provideView(context: DirectoryViewContext): ProviderResult<View>
     }
 
-    export interface ThemableFileDecorationStyle {
-        /** The CSS color property value for the text contet */
-        color?: string
-
-        /** Overwrite style for when the file is active */
-        activeColor?: string
-    }
-
-    /** A decoration attachment adds content after a {@link FileDecoration}. */
-    export interface FileDecorationAttachmentRenderOptions extends ThemableFileDecorationStyle {
-        /** Text value to be displayed. This value should be very short to prevent truncation */
-        contentText: string
-
-        /** Tooltip text to display when hovering over the text content. */
-        hoverMessage?: string
-
-        /** Overwrite style for light themes. */
-        light?: ThemableFileDecorationStyle
-
-        /** Overwrite color for dark themes. */
-        dark?: ThemableFileDecorationStyle
-    }
-
-    /**
-     * A file decoration adds text content and/or a progress bar to files in a tree view
-     */
-    export interface FileDecoration {
-        /** The resource identifier of this file */
-        uri: string
-
-        /** Whether to display the decoration on the sidebar file tree or tree page. If omitted, it will be displayed in both locations  */
-        where?: 'sidebar' | 'page'
-
-        /** An optional object that describes the text content contributed by the decoration */
-        after?: FileDecorationAttachmentRenderOptions
-
-        /**
-         * Describes a meter bar like the [HTML5 `<meter>`
-         * element](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meter)
-         * to be rendered after the file/directory name.
-         */
-        meter?: {
-            /**
-             * The current numeric value. This must be between the minimum and maximum values
-             * (min attribute and max attribute) if they are specified.
-             */
-            value: number
-
-            /**
-             * The lower numeric bound of the measured range. This must be less than
-             * the maximum value (max attribute), if specified. If unspecified, the
-             * minimum value is 0.
-             */
-            min?: number
-
-            /**
-             * The upper numeric bound of the measured range. This must be greater
-             * than the minimum value (min attribute), if specified. If unspecified,
-             * the maximum value is 1.
-             */
-            max?: number
-
-            /**
-             * The upper numeric bound of the low end of the measured range. This
-             * must be greater than the minimum value (min attribute), and it also
-             * must be less than the high value and maximum value (high attribute
-             * and max attribute, respectively), if any are specified. If
-             * unspecified, or if less than the minimum value, the low value is
-             * equal to the minimum value.
-             */
-            low?: number
-
-            /**
-             * The lower numeric bound of the high end of the measured range. This
-             * must be less than the maximum value (max attribute), and it also must
-             * be greater than the low value and minimum value (low attribute and
-             * min attribute, respectively), if any are specified. If unspecified,
-             * or if greater than the maximum value, the high value is equal to the
-             * maximum value.
-             */
-            high?: number
-
-            /**
-             * This attribute indicates the optimal numeric value. It must be within
-             * the range (as defined by the min attribute and max attribute). When
-             * used with the low attribute and high attribute, it gives an
-             * indication where along the range is considered preferable. For
-             * example, if it is between the min attribute and the low attribute,
-             * then the lower range is considered preferred. The browser may color
-             * the meter's bar differently depending on whether the value is less
-             * than or equal to the optimum value.
-             */
-            optimum?: number
-
-            /** Tooltip text to display when hovering over the progress bar. */
-            hoverMessage?: string
-        }
-    }
-
-    /**
-     * Context passed to file decoration providers.
-     *
-     * The schema of these parameters is experimental and subject to change without notice.
-     */
-    export interface FileDecorationContext {
-        /** The uri of the file's parent */
-        uri: string
-
-        files: {
-            /** The uri of the file */
-            uri: string
-
-            /** Whether this file is a directory */
-            isDirectory: boolean
-
-            /**
-             * File path relative to repo root uri
-             *
-             * @todo Remove this once `parseRepoUri` is public
-             * */
-            path: string
-        }[]
-    }
-
-    export interface FileDecorationProvider {
-        provideFileDecorations: (fileDecorationContext: FileDecorationContext) => ProviderResult<FileDecoration[]>
-    }
-
     /**
      * The client application that is running the extension.
      */
@@ -1102,14 +879,6 @@ declare module 'sourcegraph' {
         export function createPanelView(id: string): PanelView
 
         /**
-         * Creates a decorationType that can be used to add decorations to code views.
-         *
-         * Use this to create a unique handle to a set of decorations, that can be applied to
-         * text editors using {@link setDecorations}.
-         */
-        export function createDecorationType(): TextDocumentDecorationType
-
-        /**
          * Register a view provider, which provides the contents of a view.
          *
          * This API is experimental and is subject to change or removal without notice.
@@ -1119,11 +888,6 @@ declare module 'sourcegraph' {
          * @returns An unsubscribable to unregister this provider.
          */
         export function registerViewProvider(id: string, provider: ViewProvider): Unsubscribable
-
-        /**
-         * Register a file decoration provider
-         */
-        export function registerFileDecorationProvider(provider: FileDecorationProvider): Unsubscribable
 
         /**
          * Log a message to the console if logs for the extension are enabled in user settings.