+++ client/shared/gulpfile.js
@@ -49,6 +49,13 @@ async function shouldRegenerateGraphQlOperations(type, name) {
  * Generates the new query-specific types on file changes.
  */
 function watchGraphQlOperations() {
+  if (process.env.DEV_WEB_BUILDER_UNSAFE_FAST) {
+    // Setting the env var DEV_WEB_BUILDER_UNSAFE_FAST skips various operations in frontend dev.
+    // It's not safe, but if you know what you're doing, go ahead and use it. (CI will catch any
+    // issues you forgot about.)
+    return
+  }
+
   // Although graphql-codegen has watching capabilities, they don't appear to
   // use chokidar correctly and rely on polling. Instead, let's get gulp to
   // watch for us, since we know it'll do it more efficiently, and then we can