+++ client/web/src/enterprise/batches/detail/backend.ts
@@ -80,6 +80,7 @@ const bulkOperationFragment = gql`
                         url
                     }
                     repository {
+                        id
                         name
                         url
                     }