+++ client/web/src/repo/settings/RepoSettingsOptionsPage.tsx
@@ -1,22 +1,39 @@
-import { FC, useEffect } from 'react'
+import { FC, useCallback, useEffect, useState } from 'react'
 
+import { noop } from 'lodash'
 import { RouteComponentProps } from 'react-router'
 
-import { useQuery } from '@sourcegraph/http-client'
-import { Container, ErrorAlert, H2, LoadingSpinner, PageHeader, Text } from '@sourcegraph/wildcard'
+import { useMutation, useQuery } from '@sourcegraph/http-client'
+import {
+    Alert,
+    Button,
+    Container,
+    ErrorAlert,
+    H2,
+    LoadingSpinner,
+    PageHeader,
+    renderError,
+    Text,
+} from '@sourcegraph/wildcard'
 
 import { CopyableText } from '../../components/CopyableText'
-import { ExternalServiceCard } from '../../components/externalServices/ExternalServiceCard'
-import { defaultExternalServices } from '../../components/externalServices/externalServices'
 import { PageTitle } from '../../components/PageTitle'
 import {
+    ExcludeRepoFromExternalServicesResult,
+    ExcludeRepoFromExternalServicesVariables,
     SettingsAreaRepositoryFields,
     SettingsAreaRepositoryResult,
     SettingsAreaRepositoryVariables,
+    SiteExternalServiceConfigResult,
+    SiteExternalServiceConfigVariables,
 } from '../../graphql-operations'
+import { SITE_EXTERNAL_SERVICE_CONFIG } from '../../site-admin/backend'
 import { eventLogger } from '../../tracking/eventLogger'
 
-import { FETCH_SETTINGS_AREA_REPOSITORY_GQL } from './backend'
+import { EXCLUDE_REPO_FROM_EXTERNAL_SERVICES, FETCH_SETTINGS_AREA_REPOSITORY_GQL } from './backend'
+import { ExternalServiceEntry } from './components/ExternalServiceEntry'
+
+import styles from './RepoSettingsOptionsPage.module.scss'
 
 interface Props extends RouteComponentProps<{}> {
     repo: SettingsAreaRepositoryFields
@@ -25,51 +42,124 @@ interface Props extends RouteComponentProps<{}> {
 /**
  * The repository settings options page.
  */
-export const RepoSettingsOptionsPage: FC<Props> = ({ repo: propsRepo }) => {
+export const RepoSettingsOptionsPage: FC<Props> = ({ repo, history }) => {
     useEffect(() => {
         eventLogger.logViewEvent('RepoSettings')
     })
 
     const { data, error, loading } = useQuery<SettingsAreaRepositoryResult, SettingsAreaRepositoryVariables>(
         FETCH_SETTINGS_AREA_REPOSITORY_GQL,
-        { variables: { name: propsRepo.name } }
+        { variables: { name: repo.name } }
     )
 
+    // This state shows that any of possible "exclude" buttons (in this or child components) were pushed.
+    // It is used to disable all the "exclude" buttons except the button which was actually clicked.
+    const [exclusionInProgress, setExclusionInProgress] = useState<boolean>(false)
+    const [ttl, setTtl] = useState<number>(3)
+
+    // Callback used in child components (ExternalServiceEntry) to update the state in current component.
+    const updateExclusion = useCallback((updatedExclusionState: boolean) => {
+        setExclusionInProgress(updatedExclusionState)
+    }, [])
+
     const services = data?.repository?.__typename === 'Repository' && data?.repository?.externalServices.nodes
+
+    const { data: siteConfigData, error: siteConfigError } = useQuery<
+        SiteExternalServiceConfigResult,
+        SiteExternalServiceConfigVariables
+    >(SITE_EXTERNAL_SERVICE_CONFIG, {})
+
+    const [excludeRepo, { data: excludeData, error: excludeError, loading: isExcluding }] = useMutation<
+        ExcludeRepoFromExternalServicesResult,
+        ExcludeRepoFromExternalServicesVariables
+    >(EXCLUDE_REPO_FROM_EXTERNAL_SERVICES, {
+        onCompleted: () => {
+            let count = 3
+            const interval = setInterval(() => {
+                if (count === 0) {
+                    clearInterval(interval)
+                    history.push('/site-admin/external-services')
+                }
+                setTtl(count)
+                count--
+            }, 700)
+        },
+    })
+
+    const excludingDisabled =
+        (!siteConfigError &&
+            siteConfigData?.site?.externalServicesFromFile &&
+            !siteConfigData?.site?.allowEditExternalServicesWithFile) ||
+        false
 
     return (
         <>
             <PageTitle title="Repository settings" />
             <PageHeader path={[{ text: 'Settings' }]} headingElement="h2" className="mb-3" />
             <Container className="repo-settings-options-page">
+                <H2 className="mb-3">Repository name</H2>
+                <CopyableText className="mb-3" text={repo.name} size={repo.name.length} />
                 <H2 className="mb-3">Code hosts</H2>
                 {loading && <LoadingSpinner />}
                 {error && <ErrorAlert error={error} />}
                 {services && services.length > 0 && (
-                    <div className="mb-3">
+                    <div>
                         {services.map(service => (
-                            <div className="mb-3" key={service.id}>
-                                <ExternalServiceCard
-                                    {...defaultExternalServices[service.kind]}
-                                    kind={service.kind}
-                                    title={service.displayName}
-                                    shortDescription="Update this external service configuration to manage repository mirroring."
-                                    to={`/site-admin/external-services/${service.id}`}
-                                    toIcon={null}
-                                    bordered={false}
-                                />
-                            </div>
+                            <ExternalServiceEntry
+                                key={service.id}
+                                service={service}
+                                excludingDisabled={excludingDisabled}
+                                excludingLoading={exclusionInProgress}
+                                updateExclusionLoading={updateExclusion}
+                                repo={repo}
+                                history={history}
+                            />
                         ))}
                         {services.length > 1 && (
-                            <Text>
-                                This repository is mirrored by multiple external services. To change access, disable, or
-                                remove this repository, the configuration must be updated on all external services.
-                            </Text>
-                        )}
+                            <>
+                                <Text>
+                                    This repository is mirrored by multiple code hosts. To change access, disable, or
+                                    remove this repository, the configuration must be updated on all code hosts.
+                                </Text>
+                                <Button
+                                    variant="primary"
+                                    className={styles.button}
+                                    onClick={event => {
+                                        event.preventDefault()
+                                        setExclusionInProgress(true)
+                                        excludeRepo({
+                                            variables: {
+                                                externalServices: services.map(svc => svc.id),
+                                                repo: repo.id,
+                                            },
+                                        })
+                                            .catch(
+                                                // noop here is used because update error is handled directly when useMutation is called
+                                                noop
+                                            )
+                                            .finally(() => {
+                                                setExclusionInProgress(false)
+                                            })
+                                    }}
+                                    disabled={excludingDisabled || (exclusionInProgress && !isExcluding)}
+                                >
+                                    <span className={exclusionInProgress ? styles.invisibleText : ''}>
+                                        Exclude repository from all code hosts
+                                    </span>
+                                    {exclusionInProgress && <LoadingSpinner className={styles.loader} />}
+                                </Button>
+                                {excludeError && (
+                                    <ErrorAlert error={`Failed to exclude repository: ${renderError(excludeError)}`} />
+                                )}
+                                {excludeData && (
+                                    <Alert className="mt-2" variant="success">
+                                        {`Code host configurations updated. You will be redirected in ${ttl}...`}
+                                    </Alert>
+                                )}
+                            </>
+                        )}{' '}
                     </div>
                 )}
-                <H2 className="mb-3">Repository name</H2>
-                <CopyableText text={propsRepo.name} size={propsRepo.name.length} />
             </Container>
         </>
     )