+++ client/web/src/stores/experimentalFeatures.ts
@@ -8,7 +8,7 @@ const defaultSettings: SettingsExperimentalFeatures = {
     codeMonitoring: true,
     showEnterpriseHomePanels: false,
     /**
-     * Whether we show the mulitiline editor at /search/console
+     * Whether we show the multiline editor at /search/console
      */
     showMultilineSearchConsole: false,
     showSearchContext: true,