+++ internal/search/repos/repos_test.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os"
 	"reflect"
-	"sort"
 	"testing"
 	"time"
 
@@ -281,7 +280,7 @@ func BenchmarkGetRevsForMatchedRepo(b *testing.B) {
 	})
 }
 
-func TestResolverPaginate(t *testing.T) {
+func TestResolverIterator(t *testing.T) {
 	ctx := context.Background()
 	logger := logtest.Scoped(t)
 	db := database.NewDB(logger, dbtest.NewDB(logger, t))
@@ -410,13 +409,16 @@ func TestResolverPaginate(t *testing.T) {
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			r := NewResolver(logtest.Scoped(t), db, gsClient, nil, nil)
+			it := r.Iterator(ctx, tc.opts)
 
 			var pages []Resolved
-			err := r.Paginate(ctx, tc.opts, func(page *Resolved) error {
-				pages = append(pages, *page)
-				return nil
-			})
 
+			for it.Next() {
+				page := it.Current()
+				pages = append(pages, page)
+			}
+
+			err = it.Err()
 			if !errors.Is(err, tc.err) {
 				t.Errorf("%s unexpected error (-have, +want):\n%s", tc.name, cmp.Diff(err, tc.err))
 			}
@@ -426,88 +428,6 @@ func TestResolverPaginate(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestResolveRepositoriesWithUserSearchContext(t *testing.T) {
-	const (
-		wantName   = "alice"
-		wantUserID = 123
-	)
-
-	repos := database.NewMockRepoStore()
-	repos.ListMinimalReposFunc.SetDefaultHook(func(ctx context.Context, op database.ReposListOptions) ([]types.MinimalRepo, error) {
-		if op.UserID != wantUserID {
-			t.Fatalf("got %q, want %q", op.UserID, wantUserID)
-		}
-		return []types.MinimalRepo{
-			{
-				ID:   1,
-				Name: "example.com/a",
-			},
-			{
-				ID:   2,
-				Name: "example.com/b",
-			},
-			{
-				ID:   3,
-				Name: "example.com/c",
-			},
-			{
-				ID:   4,
-				Name: "external.com/a",
-			},
-			{
-				ID:   5,
-				Name: "external.com/b",
-			},
-			{
-				ID:   6,
-				Name: "external.com/c",
-			},
-		}, nil
-	})
-
-	ns := database.NewMockNamespaceStore()
-	ns.GetByNameFunc.SetDefaultHook(func(ctx context.Context, name string) (*database.Namespace, error) {
-		if name != wantName {
-			t.Fatalf("got %q, want %q", name, wantName)
-		}
-		return &database.Namespace{Name: wantName, User: wantUserID}, nil
-	})
-
-	db := database.NewMockDB()
-	db.ReposFunc.SetDefaultReturn(repos)
-	db.NamespacesFunc.SetDefaultReturn(ns)
-
-	op := search.RepoOptions{
-		SearchContextSpec: "@" + wantName,
-	}
-	repositoryResolver := NewResolver(logtest.Scoped(t), db, nil, nil, nil)
-	resolved, err := repositoryResolver.Resolve(context.Background(), op)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var got []api.RepoName
-	for _, rev := range resolved.RepoRevs {
-		got = append(got, rev.Repo.Name)
-	}
-	sort.Slice(got, func(i, j int) bool {
-		return got[i] < got[j]
-	})
-	want := []api.RepoName{
-		"example.com/a",
-		"example.com/b",
-		"example.com/c",
-		"external.com/a",
-		"external.com/b",
-		"external.com/c",
-	}
-	if diff := cmp.Diff(got, want, nil); diff != "" {
-		t.Errorf("unexpected diff: %s", diff)
-	}
-
-	mockrequire.Called(t, ns.GetByNameFunc)
-	mockrequire.Called(t, repos.ListMinimalReposFunc)
 }
 
 func TestResolveRepositoriesWithSearchContext(t *testing.T) {